Move nacl code into separate files for cleanliness, rename tweetnacl to nacl
