+++ ironfish-rust-nodejs/index.js
@@ -236,8 +236,14 @@ if (!nativeBinding) {
   throw new Error(`Failed to load native binding`)
 }
 
-const { NoteEncrypted, Note, TransactionPosted, Transaction, generateKey, generateNewPublicAddress, initializeSapling, FoundBlockResult, ThreadPoolHandler, KEY_LENGTH, NONCE_LENGTH, BoxKeyPair, randomBytes, boxMessage, unboxMessage } = nativeBinding
+const { KEY_LENGTH, NONCE_LENGTH, BoxKeyPair, randomBytes, boxMessage, unboxMessage, NoteEncrypted, Note, TransactionPosted, Transaction, generateKey, generateNewPublicAddress, initializeSapling, FoundBlockResult, ThreadPoolHandler } = nativeBinding
 
+module.exports.KEY_LENGTH = KEY_LENGTH
+module.exports.NONCE_LENGTH = NONCE_LENGTH
+module.exports.BoxKeyPair = BoxKeyPair
+module.exports.randomBytes = randomBytes
+module.exports.boxMessage = boxMessage
+module.exports.unboxMessage = unboxMessage
 module.exports.NoteEncrypted = NoteEncrypted
 module.exports.Note = Note
 module.exports.TransactionPosted = TransactionPosted
@@ -247,9 +253,3 @@ module.exports.generateNewPublicAddress = generateNewPublicAddress
 module.exports.initializeSapling = initializeSapling
 module.exports.FoundBlockResult = FoundBlockResult
 module.exports.ThreadPoolHandler = ThreadPoolHandler
-module.exports.KEY_LENGTH = KEY_LENGTH
-module.exports.NONCE_LENGTH = NONCE_LENGTH
-module.exports.BoxKeyPair = BoxKeyPair
-module.exports.randomBytes = randomBytes
-module.exports.boxMessage = boxMessage
-module.exports.unboxMessage = unboxMessage
